Split up mujoco_env into base, mujoco_env, and mujoco_py_env
Adjusted all imports
The split files contain the commit history
