+++ gymnasium/envs/mujoco/__init__.py
@@ -1,2 +1,1 @@
-from gymnasium.envs.mujoco.mujoco_env import MujocoEnv, MuJocoPyEnv
 from gymnasium.envs.mujoco.mujoco_rendering import MujocoRenderer
