+++ gymnasium/envs/mujoco/humanoid_v5.py
@@ -5,7 +5,7 @@ from typing import Dict, Tuple, Union
 import numpy as np
 
 from gymnasium import utils
-from gymnasium.envs.mujoco import MujocoEnv
+from gymnasium.envs.mujoco.mujoco_env import MujocoEnv
 from gymnasium.spaces import Box
