+++ gymnasium/envs/mujoco/inverted_pendulum.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from gymnasium import utils
-from gymnasium.envs.mujoco import MuJocoPyEnv
+from gymnasium.envs.mujoco.mujoco_py_env import MuJocoPyEnv
 from gymnasium.spaces import Box
