+++ gymnasium/envs/mujoco/reacher_v4.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from gymnasium import utils
-from gymnasium.envs.mujoco import MujocoEnv
+from gymnasium.envs.mujoco.mujoco_env import MujocoEnv
 from gymnasium.spaces import Box
